+++ packages/@uppy/utils/src/generateFileID.js
@@ -1,3 +1,5 @@
+import getFileType from './getFileType.js'
+
 function encodeCharacter (character) {
   return character.charCodeAt(0).toString(32)
 }
@@ -42,4 +44,24 @@ export default function generateFileID (file) {
   }
 
   return id
+}
+
+// If the provider has a stable, unique ID, then we can use that to identify the file.
+// Then we don't have to generate our own ID, and we can add the same file many times if needed (different path)
+function hasFileStableId (file) {
+  if (!file.isRemote || !file.remote) return false
+  // These are the providers that it seems like have stable IDs for their files. The other's I haven't checked yet.
+  const stableIdProviders = new Set(['box', 'dropbox', 'drive', 'facebook', 'unsplash'])
+  return stableIdProviders.has(file.remote.provider)
+}
+
+export function getSafeFileId (file) {
+  if (hasFileStableId(file)) return file.id
+
+  const fileType = getFileType(file)
+
+  return generateFileID({
+    ...file,
+    type: fileType,
+  })
 }
